Remove material UI expansion panel
